MAINT: Rename regulariztion folder to restoration, following naming convention from skimage.
