+++ src/darsia/__init__.py
@@ -38,9 +38,9 @@ from darsia.signals.models.dynamicthresholdmodel import *
 from darsia.signals.models.binarydataselector import *
 from darsia.signals.reduction.signalreduction import *
 from darsia.signals.reduction.monochromatic import *
-from darsia.regularization.tvd import *
-from darsia.regularization.resize import *
-from darsia.regularization.binaryinpaint import *
+from darsia.restoration.tvd import *
+from darsia.restoration.resize import *
+from darsia.restoration.binaryinpaint import *
 from darsia.analysis.translationanalysis import *
 from darsia.analysis.concentrationanalysis import *
 from darsia.analysis.compactionanalysis import *
