+++ src/darsia/restoration/binaryinpaint.py
